Tests: Rename 'tests/' files w/ correct case
