+++ src/artist/artist.controller.spec.ts
@@ -6,9 +6,9 @@ import FileManagerService from "src/file-manager/file-manager.service";
 import PrismaModule from "src/prisma/prisma.module";
 import PrismaService from "src/prisma/prisma.service";
 import SongModule from "src/song/song.module";
-import { FakeFileManagerService } from "test/FakeFileManagerModule";
-import SetupApp from "test/SetupApp";
-import { createTestingModule } from "test/TestModule";
+import { FakeFileManagerService } from "test/fake-file-manager.module";
+import SetupApp from "test/setup-app";
+import { createTestingModule } from "test/test-module";
 import ArtistModule from "./artist.module";
 import ArtistService from "./artist.service";
 import request from 'supertest';
