+++ src/file/file.controller.spec.ts
@@ -6,9 +6,9 @@ import LibraryModule from "src/library/library.module";
 import LibraryService from "src/library/library.service";
 import PrismaModule from "src/prisma/prisma.module";
 import PrismaService from "src/prisma/prisma.service";
-import { FakeFileManagerService } from "test/FakeFileManagerModule";
-import SetupApp from "test/SetupApp";
-import { createTestingModule } from "test/TestModule";
+import { FakeFileManagerService } from "test/fake-file-manager.module";
+import SetupApp from "test/setup-app";
+import { createTestingModule } from "test/test-module";
 import FileModule from "./file.module";
 import FileService from "./file.service";
 import request from 'supertest';
