+++ src/file/file.service.spec.ts
@@ -15,8 +15,8 @@ import ReleaseModule from "src/release/release.module";
 import SettingsModule from "src/settings/settings.module";
 import SongModule from "src/song/song.module";
 import TrackModule from "src/track/track.module";
-import { FakeFileManagerService } from "test/FakeFileManagerModule";
-import { createTestingModule } from "test/TestModule";
+import { FakeFileManagerService } from "test/fake-file-manager.module";
+import { createTestingModule } from "test/test-module";
 import { FileAlreadyExistsException, FileNotFoundFromIDException } from "./file.exceptions";
 import FileModule from "./file.module";
 import FileService from "./file.service";
