+++ src/genre/genre.controller.spec.ts
@@ -13,8 +13,8 @@ import PrismaModule from "src/prisma/prisma.module";
 import PrismaService from "src/prisma/prisma.service";
 import SongService from "src/song/song.service";
 import TrackModule from "src/track/track.module";
-import { FakeFileManagerService } from "test/FakeFileManagerModule";
-import { createTestingModule } from "test/TestModule";
+import { FakeFileManagerService } from "test/fake-file-manager.module";
+import { createTestingModule } from "test/test-module";
 import GenreModule from "./genre.module";
 import GenreService from "./genre.service";
 import MetadataModule from "src/metadata/metadata.module";
