+++ src/genre/genre.service.spec.ts
@@ -10,8 +10,8 @@ import PrismaService from "src/prisma/prisma.service";
 import Slug from "src/slug/slug";
 import SongService from "src/song/song.service";
 import TrackModule from "src/track/track.module";
-import { FakeFileManagerService } from "test/FakeFileManagerModule";
-import { createTestingModule } from "test/TestModule";
+import { FakeFileManagerService } from "test/fake-file-manager.module";
+import { createTestingModule } from "test/test-module";
 import { GenreAlreadyExistsException, GenreNotFoundByIdException, GenreNotFoundException } from "./genre.exceptions";
 import GenreModule from "./genre.module";
 import GenreService from "./genre.service";
