+++ src/library/library.service.spec.ts
@@ -1,9 +1,9 @@
-import { createTestingModule } from "test/TestModule";
+import { createTestingModule } from "test/test-module";
 import type { TestingModule } from "@nestjs/testing";
 import FileManagerService from "src/file-manager/file-manager.service";
 import PrismaService from "src/prisma/prisma.service";
 import Slug from "src/slug/slug";
-import { FakeFileManagerService } from "test/FakeFileManagerModule";
+import { FakeFileManagerService } from "test/fake-file-manager.module";
 import { LibraryAlreadyExistsException, LibraryNotFoundException } from "./library.exceptions";
 import LibraryService from "./library.service";
 import LibraryModule from "./library.module";
