+++ src/metadata/metadata.service.spec.ts
@@ -8,8 +8,8 @@ import ReleaseModule from "src/release/release.module";
 import SettingsModule from "src/settings/settings.module";
 import SongModule from "src/song/song.module";
 import TrackModule from "src/track/track.module";
-import { FakeFileManagerService } from "test/FakeFileManagerModule";
-import { createTestingModule } from "test/TestModule";
+import { FakeFileManagerService } from "test/fake-file-manager.module";
+import { createTestingModule } from "test/test-module";
 import { PathParsingException } from "./metadata.exceptions";
 import MetadataModule from "./metadata.module";
 import MetadataService from "./metadata.service";
